requirements.txt
streamlit==1.32.0
groq==0.5.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
pypdf==4.1.0
langchain==0.1.16
langchain-community==0.0.32
tiktoken==0.6.0
